Moved check_senders2 to check_senders
